asr_test_files/model_4/requirements.txt
transformers==4.31.0
librosa==0.10.1
fastapi
accelerate
uvicorn[standard]
pandas==2.0.3
openai-whisper
librosa
noisereduce